+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.18] - 2026-01-22
+
+### Added
+- **Function serialization** in `SimpleRedisCheckpointer`
+  - Functions/callables serialized as `{"__type__": "function", "value": null}`
+  - Allows LangGraph internals that include callables to be checkpointed
+  - 3 new unit tests for function serialization
+
 ## [0.3.17] - 2026-01-22
 
 ### Added